Commit ver 1.84
Fixed the issue where the player is out of range but the notice component is still registering the player as in. I changed it by moving the function of notice component in late update. HOWEVER, there is a bug where the zombie does not stop moving after not having a target.